реализована валидация в UI на странице входа и регистрации
